requirements.txt
flask
flask_cors
torch
sentence_transformers
openai
python-dotenv
waitress